Module dslab_storage::disk

Basic implementation of storage model for disk.

It supports modeling of concurrent execution of disk operations by means of generic fair throughput sharing model from the dslab-models crate. The underlying model also supports modeling of throughput degradation, variability and dependence on operation properties by means of user-defined throughput and factor functions. For detailed information about these functions, please refer to documentation in dslab-models crate.

Note that this model is quite generic and can be used to model other types of storage as well.

Structs

Disk

Represents a disk.

DiskBuilder

Disk builder. This is a type for convenient disk setup.

DiskOperation

Describes a disk operation.

Enums

DiskOperationType

Type of disk operation.