Struct dslab_storage::fs::FileSystem

pub struct FileSystem { /* private fields */ }

Representation of file system.

Implementations

impl FileSystem

pub fn new(ctx: SimulationContext) -> Self

Creates new empty file system.

pub fn mount_disk( &mut self, mount_point: &str, disk: Rc<RefCell<dyn Storage>> ) -> Result<(), String>

Mounts disk to mount_point if it is not taken yet.

pub fn unmount_disk(&mut self, mount_point: &str) -> Result<(), String>

Unmounts a disk which is mounted to mount_point if there is any.

pub fn read(&mut self, file_path: &str, size: u64, requester: Id) -> u64

Submits file read request and returns unique request id.

The amount of data read from file located at file_path is specified in size. The component specified in requester will receive FileReadCompleted event upon the read completion. If the read size is larger than the file size, FileReadFailed event will be immediately emitted instead. Note that the returned request id is unique only within the current file system.

pub fn read_all(&mut self, file_path: &str, requester: Id) -> u64

Submits file read request and returns unique request id.

The amount of data read from file located at file_path is equal to the file size. The component specified in requester will receive FileReadCompleted event upon the read completion. Note that the returned request id is unique only within the current file system.

pub fn write(&mut self, file_path: &str, size: u64, requester: Id) -> u64

Submits file write request and returns unique request id.

The amount of data written to file located at file_path is specified in size. The component specified in requester will receive FileWriteCompleted event upon the write completion. If there is not enough available disk space, FileWriteFailed event will be immediately emitted instead. Note that the returned request id is unique only within the current file system.

pub fn create_file(&mut self, file_path: &str) -> Result<(), String>

Creates file at file_path if it doesn’t already exist.

pub fn file_size(&self, file_path: &str) -> Result<u64, String>

Returns size of the file located at file_path if there is any.

pub fn used_space(&self) -> u64

Returns amount of used space on all disks currently mounted to this file system.

pub fn free_space(&self) -> u64

Returns amount of free space on all disks currently mounted to this file system.

pub fn capacity(&self) -> u64

Returns cumulative capacity of all disks currently mounted to this file system.

pub fn disks_info(&self) -> Vec<(String, StorageInfo)>

Returns vec of disk info associated with mount points.

pub fn disk_info(&self, mount_point: &str) -> Result<StorageInfo, String>

Returns disk info for a mount point.

pub fn mount_points(&self) -> Vec<String>

Returns mount points present in this file system.

pub fn delete_file(&mut self, file_path: &str) -> Result<(), String>

Deletes file located at file_path if there is any.