Trait dslab_storage::storage::Storage

pub trait Storage {
    // Required methods
    fn read(&mut self, size: u64, requester: Id) -> u64;
    fn write(&mut self, size: u64, requester: Id) -> u64;
    fn mark_free(&mut self, size: u64) -> Result<(), String>;
    fn used_space(&self) -> u64;
    fn free_space(&self) -> u64;
    fn capacity(&self) -> u64;
    fn id(&self) -> Id;
    fn info(&self) -> StorageInfo;
}

A trait for modeling an abstract storage resource, i.e. any device, system or service for storing data.

The main operations are reading and writing data to/from storage. Implementations should model the delays associated with these operations. The trait also includes operations to get information about the storage, including its capacity, current usage, etc.

Required Methods

fn read(&mut self, size: u64, requester: Id) -> u64

Submits data read request and returns unique request id.

The amount of data read from storage is specified in size. The component specified in requester will receive DataReadCompleted event upon the read completion. If the read size is larger than the storage capacity, DataReadFailed event will be immediately emitted instead. Note that the returned request id is unique only within the current storage.

fn write(&mut self, size: u64, requester: Id) -> u64

Submits data write request and returns unique request id.

The amount of data written to storage is specified in size. The component specified in requester will receive DataWriteCompleted event upon the write completion. If there is not enough available storage space, DataWriteFailed event will be immediately emitted instead. Note that the returned request id is unique only within the current storage.

fn mark_free(&mut self, size: u64) -> Result<(), String>

Marks previously used storage space of given size as free.

The size should not exceed the currently used storage space.

fn used_space(&self) -> u64

Returns the amount of used storage space.

fn free_space(&self) -> u64

Returns the amount of free storage space.

fn capacity(&self) -> u64

Returns the storage capacity.

fn id(&self) -> Id

Returns identifier of simulation component representing the storage.

fn info(&self) -> StorageInfo

Returns struct with information about the storage.